Themes: Split theme sheet content into separate component
